Add files via upload
XYZ Pre-Primary School was established in 2007.This school was famous for high quality education in pre-primary classes. Now, School has experienced varying enrollment rates over the years. Understanding the underlying factors and addressing the challenges is crucial for sustained growth. This report aims to analyse the enrolment trends and provide actionable strategies to enhance enrolments.